Requisitos completos

+++ src/main/java/com/samuelsoares/client/services/ClientService.java
@@ -1,14 +1,20 @@
 package com.samuelsoares.client.services;
 
-import java.util.List;
-
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.samuelsoares.client.dto.ClientDTO;
 import com.samuelsoares.client.entities.Client;
 import com.samuelsoares.client.repositories.ClientRepository;
+import com.samuelsoares.client.services.exceptions.DatabaseException;
+import com.samuelsoares.client.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ClientService {
@@ -18,14 +24,16 @@ public class ClientService {
 	
 	@Transactional(readOnly = true)
 	public ClientDTO findById(Long id) {
-		Client client = repository.findById(id).get();
+		Client client = repository.findById(id).orElseThrow(
+				() -> new ResourceNotFoundException("Recurso não encontrado"));
+				
 		return new ClientDTO(client);
 	}
 	
 	@Transactional(readOnly = true)
-	public List<ClientDTO> findAll() {
-		List<Client> client = repository.findAll();
-		return client.stream().map(x -> new ClientDTO(x)).toList();
+	public Page<ClientDTO> findAll(Pageable pageable) {
+		Page<Client> result = repository.findAll(pageable);
+		return result.map(x -> new ClientDTO(x));
 	}
 	
 	@Transactional
@@ -38,14 +46,28 @@ public class ClientService {
 
 	@Transactional
 	public ClientDTO update(Long id, ClientDTO dto) {
+		try {
 		Client entity = repository.getReferenceById(id);
 		copyDToToEntity(dto, entity);
 		entity = repository.save(entity);
 		return new ClientDTO(entity);
+		}
+		catch(EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
 	}
 	
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		repository.deleteById(id);
+		if (!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		try {
+	        	repository.deleteById(id);    		
+		}
+	    	catch (DataIntegrityViolationException e) {
+	        	throw new DatabaseException("Falha de integridade referencial");
+	   	}
 	}
 	
 	private void copyDToToEntity(ClientDTO dto, Client entity) {
